### 2024-10-08 #7331 @editorialbot
[PRE REVIEW]: Snk: A Snakemake CLI and Workflow Management System
**Submitting author:** @wytamma (<a href="http://orcid.org/0000-0001-7070-0078">Wytamma Wirth</a>)
**Repository:** https://github.com/Wytamma/snk
**Branch with paper.md** (empty if default branch):
**Version:** v0.30.1
