### 2024-10-08 #7331 @Wytamma
Thanks @danielskatz,

Just FYI the majority of the code is in the snk sister project [Snk-CLI](https://github.com/Wytamma/snk-cli). There are two different projects so that users can use Snk-CLI without depending on the main Snk application. Details here -> https://snk.wytamma.com/workflow_packages and in the paper.md.

Kind regards,

Wytamma
