### 2024-10-08 dc350ab
Update README.md
:abc: adding JOSS badge
