+++ README.md
@@ -19,6 +19,9 @@ Snk (pronounced snek) is a Snakemake workflow management system. Snk allows you 
 ## Installation
 
 ```console
+conda install bioconda::snk
+```
+```console
 pip install snk
 ```
 
