### 2024-10-24 #7331 @Wytamma
Thanks @danielskatz!

I've added potential reviewers here: [beardymcjohnface](https://reviewers.joss.theoj.org/reviewers/3486), [cmeesters](https://reviewers.joss.theoj.org/reviewers/4161), [kdm9](https://reviewers.joss.theoj.org/reviewers/3337), [huddlej](https://github.com/huddlej) & [miparedes](https://github.com/miparedes).
