### 2024-10-25 #7410 @editorialbot
[REVIEW]: Snk: A Snakemake CLI and Workflow Management System
**Submitting author:** @wytamma (<a href="http://orcid.org/0000-0001-7070-0078">Wytamma Wirth</a>)
**Repository:** https://github.com/Wytamma/snk
**Branch with paper.md** (empty if default branch):
**Version:** v0.31.0
