### 2024-10-28 d2fa734
:bug: better tag error reporting. Solves #80
