### 2024-10-31 Wytamma/snk#83 @huddlej
JOSS review: Should @ktmeaton be included in the author list?
When reviewing the JOSS review checklist item about authorship, I noticed that @ktmeaton has a [nonzero code/doc contribution to snk](https://github.com/Wytamma/snk/graphs/contributors) but isn't in the author list. Should they be included since another person who does not appear to have code/manuscript contributions is listed as a co-author?

Related to https://github.com/openjournals/joss-reviews/issues/7410
