### 2024-10-31 Wytamma/snk#83 @Wytamma
Probably! @ktmeaton want to be a coauthor?

The other author was a major contributor of the [McCoy](https://github.com/mccoy-devs/mccoy/graphs/contributors) project in which Snk was initially conceived and developed.
