### 2024-10-31 340346a
Fix typo in test command
