### 2024-10-31 Wytamma/snk#83 @ktmeaton
Congratulations on your paper submission @Wytamma, `snk` is an awesome tool! I don't think my very small pull request warrants co-author status but thank you for asking 😊 (I actually didn't remember what my code contribution was until I dug through git. Which I think is additional evidence that I haven't made a substantial contribution.)
