### 2024-11-01 Wytamma/snk#85 @Wytamma
Solved in #86
