### 2024-11-01 Wytamma/snk#83 @Wytamma
Thanks @ktmeaton! I'll make sure I update the paper to include an acknowledgement of your contribution. You are a fantastic open-source community member :)
