### 2024-11-01 db5eaea
:tada: add acknowledgments
