### 2024-11-01 Wytamma/snk#88 @huddlej
JOSS review: Prompt at uninstall incorrectly suggests default behavior is to confirm
## Description

I've been trying out snk with [a simple Nextstrain workflow](https://github.com/nextstrain/zika-tutorial/pull/16) which I install and run like this:

```
# Install the workflow.
snk install \
    --resource data \
    --resource config \
    --tag add-conda-env \
    nextstrain/zika-tutorial

# Run the workflow.
zika-tutorial run
```

This works really well! When I tried uninstalling the workflow, I got the following output:

```bash
$ snk uninstall zika-tutorial
Uninstalling zika-tutorial
  Would remove:
    /Users/jlhudd/miniconda3/envs/snk/snk/workflows/zika-tutorial/*
    /Users/jlhudd/miniconda3/envs/snk/snk/bin/zika-tutorial
    /Users/jlhudd/miniconda3/envs/snk/bin/zika-tutorial
Proceed (Y/n)?
```

The capitalized "Y" in the final prompt suggests that the default behavior when I hit enter will be to confirm the uninstall. Instead, when I hit enter, nothing happens.

What if instead of capitalizing the "y" the prompt used a lowercase "y" that doesn't have the same connotation of a default behavior? Or alternately, would it be just as easy to make the default behavior match the confirmation with "y"?

Motivated by https://github.com/openjournals/joss-reviews/issues/7410
