### 2024-11-01 Wytamma/snk#89 @huddlej
JOSS review: Consider including graphviz as dependency for installation documentation
## Description

I really like that the `--dag` argument provides a simple wrapper around Snakemake’s own dot output. As I worked through [the part of the Basic Use docs that show how to build a DAG](https://snk.wytamma.com/#create-a-dag) with this argument, I had to install graphviz into the `snk` environment I had created.

This is a pretty minor (non-blocking) user experience point, but what if the graphviz dependency was included in the initial installation command in the docs like the following?

```
conda create -n snk bioconda::snk conda-forge:graphviz
```

Alternately, you could make graphviz a dependency of the snk Bioconda recipe and keep the docs in their current simpler form.

Motivated by https://github.com/openjournals/joss-reviews/issues/7410
