### 2024-11-01 Wytamma/snk#90 @huddlej
JOSS review: Clarify intended users for `snk.yaml` in documentation
## Description

[The Snk configuration file documentation](https://snk.wytamma.com/snk_config_file/) is helpfully detailed about how to create these files. I couldn’t tell whether the users referred to by this documentation were the workflow developers or the workflow runners.

As a workflow developer, I could guess that I need to place a `snk.yaml` file in the top-level of my workflow directory. As a workflow runner, I couldn’t tell where I would put a `snk.yaml` file or if should edit the `snk.yaml` that gets created automatically for a workflow that I’ve installed.

I assume from the context, that this documentation is for workflow developers, but could you clarify in the introductory text who the primary users are for the `snk.yaml` file?

Motivated by https://github.com/openjournals/joss-reviews/issues/7410
