### 2024-11-01 #7410 @huddlej
@danielskatz @Wytamma I'm finished reviewing the paper and software. It was a pleasure to try out this new tool. There are many times in the last decade that I wish I had had a tool like this; it would have made life easier for me and users of my workflows!

I've opened a few minor issues in the project's repo (linked above) which I'm happy to chat with you about, @Wytamma, if you like.
