### 2024-11-01 #7410 @danielskatz
Thanks @huddlej - I assume from the fact that you checked everything off on your list that these issues are not ones that you think need to be addressed for the submission to be published? If this is incorrect, please help me understand which issues do need to be addressed before publication.
