### 2024-11-01 #7410 @huddlej
@danielskatz Thank you for checking! I would like the open issues to be addressed before publication, even if they are addressed with a comment that they are out of scope. I'll leave the "Functionality documentation" item unchecked in my list above until then.
