### 2024-11-07 Wytamma/snk#91 @beardymcjohnface
JOSS Review: nested config
https://github.com/openjournals/joss-reviews/issues/7410

I really like the way snk attempts to handle nested config. However, I think there are still cases where users would run into problems. Example, the mapache pipeline (github.com/sneuensc/mapache) config lets the user list multiple genomes:

```yaml
# (e.g., ind1.hg19.bam)
genome:
    hg19: /work/FAC/FBM/DBC/amalaspi/popgen/shared_resources/reference_human/hs.build37.1/hs.build37.1.fa
    #test: /users/sneuensc/work_sneuensc/mapping_pipeline/mapache/test_genome.fa
    #GRCh38: /work/FAC/FBM/DBC/amalaspi/popgen/shared_resources/reference_human/GCA_000001405.15_GRCh38.p13_onlyChrom/GCA_000001405.15_GRCh38.fa
```

However, this translates into `--genome-hg19` and expects the text filepath for hg19, and it doesn't look like there is a way to use the CLI to add more genomes. I can't think of a fix that snk could implement to address this (you might be able to but it's not worth the time and effort involved), but it might be helpful to highlight that snk might not work for some config files.
