### 2024-11-07 #7410 @beardymcjohnface
This package and the idea of the snk.yaml is exactly what I've wanted to make for a while now, so I'm really happy that someone else did it and I don't have to! I would love to see this evolve into something like nf-core but for snakemake.
