### 2024-11-07 Wytamma/snk#92 @beardymcjohnface
JOSS Review: snk.yaml schema versions
https://github.com/openjournals/joss-reviews/issues/7410

The snk.yaml expects certain fields etc when building the CLI. The schema might change over time. Would it be sensible to add a schema version to the snk.yaml to ensure backwards compatibility for newer versions of snk with older pipelines using the earlier schemas? Something like:

```yaml
# snk.yaml
schema: 1.0
logo: yeet
...
```
