### 2024-11-16 Wytamma/snk#88 @Wytamma
Thanks @huddlej! I've updated the default behaviour to match confirmation.
