### 2024-11-16 Wytamma/snk#89 @Wytamma
Great Idea @huddlej! I've added graphviz as a dependancy on conda (https://github.com/bioconda/bioconda-recipes/pull/52143).
