### 2024-11-16 4a915c6
:sparkles: add snk edit command
