### 2024-11-16 40a3360
:white_check_mark: add snk edit tests
