### 2024-11-16 d3fcfb8
:sparkles: add snk create command
