### 2024-11-16 Wytamma/snk#90 @Wytamma
This is a great point @huddlej! I have really been thinking about `snk.yaml` from the workflow developer side but but makes sense for users to be able to modify it easily.

To enable this I've added a `snk edit [WORKFLOW_NAME]` command that lets users modify the `snk.yaml` config file for any installed workflow (#93). I've also added docs and tests for this feature.
