### 2024-11-16 Wytamma/snk#91 @Wytamma
Hi @beardymcjohnface, thanks for the review and great issue!

I think this will need to be solved with the `snk.yaml` file e.g. some special case like [`choices`](https://snk.wytamma.com/snk_config_file/#cli).

I think the best way to do it will be to add a [`Tuple`](https://typer.tiangolo.com/tutorial/multiple-values/options-with-multiple-values/) type and let user specify the the key value pairs with `--genome <TEXT TEXT>....`. I'm not sure if that will work with multiple values but I'll test that and get back to you.
