### 2024-11-17 775cc16
:memo: update docs for pair and dict
