+++ snk/__about__.py
@@ -1,4 +1,4 @@
 # SPDX-FileCopyrightText: 2022-present wytamma <wytamma.wirth@me.com>
 #
 # SPDX-License-Identifier: MIT
-__version__ = "0.30.1"
+__version__ = "0.31.0"
