### 2024-11-17 ef97429
:arrow_up: update snk-cli dep
