### 2024-11-17 Wytamma/snk#92 @Wytamma
Thanks again @beardymcjohnface!

I've been thinking about this one... As the scheme is tied to the `snk-cli` version and independent of `snk` it acts as a sort of schema versioning i.e. you can specify the `snk-cli` version at install time to install older pipelines or .

E.g. this will install mapache ensuring the `snk-cli` version (schema) is >= 0.6.1.

```bash
snk install https://github.com/sneuensc/mapache -d "snk_cli>=0.6.1"
```

There may be a cleaner way to do this i.e. the same way I check for min snakemake version so I'll leave this issue open for now.
