### 2024-11-17 Wytamma/snk#92 @Wytamma
I ended up just adding `min_snk_cli_version` to the SnkConfig class so you can set this in the `snk.yaml` file and snk will make sure it uses `snk_cli>=min_snk_cli_version`.
