### 2024-11-17 93ae297
:sparkles: min_snk_cli_version
