+++ pyproject.toml
@@ -25,7 +25,7 @@ classifiers = [
 ]
 dependencies = [
   "GitPython~=3.1.0",
-  "snk-cli>=0.6.1",
+  "snk-cli>=0.7.0",
 ]
 dynamic = ["version"]
 
