### 2024-11-17 6209306
:bug: allow snk-cli version to be over written
