### 2024-11-17 22a6079
:art: Update snk_config_file.md
