### 2024-11-18 Wytamma/snk#90 @huddlej
Cool! That's a great solution, @Wytamma. Thank you!
