### 2024-11-18 #7410 @danielskatz
👋 @Wytamma - At this point could you:

I can then move forward with proofreading and then accepting the submission. There's a reasonable chance I will ask for changes, but they do not need to be included in the archived software, as the paper itself will be archived as well.
