### 2024-11-24 #7410 @danielskatz
👋 @Wytamma - My suggested changes are in https://github.com/Wytamma/snk/pull/95 - please merge this before the 4 steps mentioned earlier, or let me know what you disagree with.
