### 2024-11-25 #7410 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
