### 2024-11-25 6069b34
:art: minor updates to JOSS paper
